Se actualiza los modelos, se crea una nueva carpeta para almacenar la gestion de los modelos y se actualiza el DiagramaDeClases.puml.

+++ src/main/java/co/edu/uniquindio/proyectofinal/modelos/Usuario.java
@@ -13,9 +13,9 @@ import lombok.*;
 public class Usuario {
 
     @EqualsAndHashCode.Include
-    protected int cedula;
+    protected long cedula;
 
     protected String nombreCompleto, email, contraseña;
-    protected int telefono;
+    protected long telefono;
 
 }
